Revert "Move the funcs to cairo1-run and make them public"
This reverts commit a676d0bbc14c1d105a99eb11b40b7658ae4a8990.
